Package restructure for spring-graphql
